Fix up tests
