+++ executor/tests/execute_comparison_check.rs
@@ -129,7 +129,7 @@ fn attribute_equality() {
     let (type_manager, thing_manager) = load_managers(storage.clone(), None);
     let variable_registry = &translation_context.variable_registry;
     let previous_stage_variable_annotations = &BTreeMap::new();
-    let entry_annotations = infer_types(
+    let block_annotations = infer_types(
         &snapshot,
         &entry,
         variable_registry,
@@ -139,6 +139,7 @@ fn attribute_equality() {
         false,
     )
     .unwrap();
+    let entry_annotations = block_annotations.type_annotations_of(entry.conjunction()).unwrap();
 
     let (row_vars, variable_positions, mapping, named_variables) =
         position_mapping([var_age_a, var_age_b], [var_age_type_a, var_age_type_b]);
